backend/requirements.dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.dev.in
#
amqp==2.6.1
    # via kombu
apipkg==1.5
    # via execnet
asgiref==3.3.1
    # via django
attrs==20.3.0
    # via
    #   flake8-bugbear
    #   jsonschema
    #   pytest
bandit==1.7.0
    # via flake8-bandit
billiard==3.6.3.0
    # via celery
boto3==1.17.2
    # via django-amazon-ses
botocore==1.20.2
    # via
    #   boto3
    #   s3transfer
cachetools==4.2.1
    # via premailer
celery==4.4.6
    # via -r requirements.in
certifi==2020.12.5
    # via
    #   requests
    #   sentry-sdk
chardet==4.0.0
    # via requests
colorama==0.4.4
    # via flakehell
coverage==5.4
    # via pytest-cov
cssselect==1.1.0
    # via premailer
cssutils==1.0.2
    # via premailer
dj-database-url==0.5.0
    # via -r requirements.in
django-amazon-ses==4.0.0
    # via -r requirements.in
django-ckeditor==6.0.0
    # via -r requirements.in
django-cleanup==5.1.0
    # via -r requirements.in
django-cors-headers==3.7.0
    # via -r requirements.in
django-js-asset==1.2.2
    # via django-ckeditor
django-solo==1.1.5
    # via -r requirements.in
django-storages==1.11.1
    # via -r requirements.in
django-stubs==1.7.0
    # via -r requirements.dev.in
django-tabbed-admin==1.0.4
    # via -r requirements.in
django-watchman==1.2.0
    # via -r requirements.in
django==3.1.6
    # via
    #   -r requirements.in
    #   django-amazon-ses
    #   django-cors-headers
    #   django-storages
    #   django-stubs
    #   django-watchman
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
    #   model-bakery
djangorestframework-simplejwt==4.6.0
    # via -r requirements.in
djangorestframework==3.12.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
drf-spectacular==0.13.1
    # via -r requirements.in
entrypoints==0.3
    # via flakehell
execnet==1.8.0
    # via pytest-xdist
flake8-bandit==2.1.2
    # via -r requirements.dev.in
flake8-blind-except==0.2.0
    # via -r requirements.dev.in
flake8-bugbear==20.11.1
    # via -r requirements.dev.in
flake8-builtins==1.5.3
    # via -r requirements.dev.in
flake8-comprehensions==3.3.1
    # via -r requirements.dev.in
flake8-django==1.1.1
    # via -r requirements.dev.in
flake8-functions==0.0.5
    # via -r requirements.dev.in
flake8-import-order==0.18.1
    # via -r requirements.dev.in
flake8-logging-format==0.6.0
    # via -r requirements.dev.in
flake8-polyfill==1.0.2
    # via
    #   flake8-bandit
    #   pep8-naming
flake8-print==4.0.0
    # via -r requirements.dev.in
flake8-type-annotations==0.1.0
    # via -r requirements.dev.in
flake8-variables-names==0.0.4
    # via -r requirements.dev.in
flake8==3.8.4
    # via
    #   -r requirements.dev.in
    #   flake8-bandit
    #   flake8-bugbear
    #   flake8-builtins
    #   flake8-comprehensions
    #   flake8-django
    #   flake8-polyfill
    #   flake8-print
    #   flake8-type-annotations
    #   flakehell
flakehell==0.9.0
    # via -r requirements.dev.in
future==0.18.2
    # via celery
gitdb==4.0.5
    # via gitpython
gitpython==3.1.12
    # via bandit
gunicorn==20.0.4
    # via -r requirements.in
idna==2.10
    # via requests
inflection==0.5.1
    # via drf-spectacular
iniconfig==1.1.1
    # via pytest
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsonschema==3.2.0
    # via drf-spectacular
kombu==4.6.11
    # via celery
lxml==4.6.2
    # via premailer
mccabe==0.6.1
    # via flake8
model-bakery==1.2.1
    # via -r requirements.in
mr-proper==0.0.6
    # via flake8-functions
mypy-extensions==0.4.3
    # via mypy
mypy==0.800
    # via
    #   -r requirements.dev.in
    #   django-stubs
packaging==20.9
    # via pytest
pbr==5.5.1
    # via stevedore
pep8-naming==0.11.1
    # via -r requirements.dev.in
pillow==8.1.0
    # via -r requirements.in
pluggy==0.13.1
    # via pytest
premailer==3.7.0
    # via -r requirements.in
psycopg2-binary==2.8.6
    # via -r requirements.in
py==1.10.0
    # via
    #   pytest
    #   pytest-forked
pycodestyle==2.6.0
    # via
    #   flake8
    #   flake8-bandit
    #   flake8-import-order
    #   flake8-print
pyflakes==2.2.0
    # via flake8
pygments==2.7.4
    # via flakehell
pyjwt==2.0.1
    # via djangorestframework-simplejwt
pyparsing==2.4.7
    # via packaging
pyrsistent==0.17.3
    # via jsonschema
pytest-cov==2.11.1
    # via -r requirements.in
pytest-django==4.1.0
    # via -r requirements.in
pytest-forked==1.3.0
    # via pytest-xdist
pytest-xdist==2.2.0
    # via -r requirements.in
pytest==6.2.2
    # via
    #   -r requirements.in
    #   flake8-functions
    #   pytest-cov
    #   pytest-django
    #   pytest-forked
    #   pytest-xdist
python-dateutil==2.8.1
    # via botocore
pytz==2021.1
    # via
    #   celery
    #   django
pyyaml==5.4.1
    # via
    #   bandit
    #   drf-spectacular
redis==3.5.3
    # via -r requirements.in
requests==2.25.1
    # via
    #   premailer
    #   responses
responses==0.12.1
    # via -r requirements.in
s3transfer==0.3.4
    # via boto3
sentry-sdk==0.20.3
    # via -r requirements.in
six==1.15.0
    # via
    #   bandit
    #   flake8-print
    #   jsonschema
    #   python-dateutil
    #   responses
smmap==3.0.5
    # via gitdb
sqlparse==0.4.1
    # via django
stdlib-list==0.8.0
    # via mr-proper
stevedore==3.3.0
    # via bandit
toml==0.10.2
    # via
    #   flakehell
    #   pytest
typed-ast==1.4.2
    # via mypy
typing-extensions==3.7.4.3
    # via
    #   -r requirements.in
    #   django-stubs
    #   mypy
uritemplate==3.0.1
    # via drf-spectacular
urllib3==1.26.3
    # via
    #   botocore
    #   flakehell
    #   requests
    #   responses
    #   sentry-sdk
vine==1.3.0
    # via
    #   amqp
    #   celery

# The following packages are considered to be unsafe in a requirements file:
# setuptools
